+++ src/screens/MailThreadScreen/mock-data/emails.ts
@@ -1,0 +1,72 @@
+export const emailThread = {
+  subject: 'Sahaj: Coding Assignment - UI',
+  emails: [
+    {
+      from: {
+        email: 'sahaj.ai@gmail.com',
+        name: 'Avinash',
+      },
+      to: {
+        email: 'ashim@gmail.com',
+        name: 'Ashim',
+      },
+      cc: [
+        {
+          email: 'someone@gmail.com',
+          name: 'Someone CCed',
+        },
+        {
+          email: 'someone2@gmail.com',
+          name: 'Someone 2CCed',
+        }
+      ],
+      date: '20-02-2021',
+      message: 'Hello worls, this is my email',
+    },
+    {
+      from: {
+        email: 'sahaj.ai@gmail.com',
+        name: 'Avinash',
+      },
+      to: {
+        email: 'ashim@gmail.com',
+        name: 'Ashim',
+      },
+      cc: [
+        {
+          email: 'someone@gmail.com',
+          name: 'Someone CCed',
+        },
+        {
+          email: 'someone2@gmail.com',
+          name: 'Someone 2CCed',
+        }
+      ],
+      date: '20-02-2021',
+      message: 'Hello worls, this is my email',
+    },
+    {
+      from: {
+        email: 'sahaj.ai@gmail.com',
+        name: 'Avinash',
+      },
+      to: {
+        email: 'ashim@gmail.com',
+        name: 'Ashim',
+      },
+      cc: [
+        {
+          email: 'someone@gmail.com',
+          name: 'Someone CCed',
+        },
+        {
+          email: 'someone2@gmail.com',
+          name: 'Someone 2CCed',
+        }
+      ],
+      date: '20-02-2021',
+      message: 'Hello worls, this is my email',
+    }
+  ]
+}
+
