+++ src/screens/index.ts
@@ -1,1 +1,2 @@
 export { MailboxScreen } from './MailboxScreen/MailboxScreen';
+export { MailThreadScreen } from './MailThreadScreen/MailThreadScreen';
